refactor: move getPropertyOpValues to core package
