+++ packages/core/package.json
@@ -40,7 +40,6 @@
         "ts-node": "10.1.0",
         "typescript": "4.4.3",
         "typescript-standard": "^0.3.36",
-        "wot-thing-description-types": "1.1.0-23-March-2023",
         "wot-typescript-definitions": "0.8.0-SNAPSHOT.25"
     },
     "dependencies": {
